Boost enchantment on randart plates by a lot, slightly nerf robes.
Most heavy randarts being +0..+3, with an absolute max of +6, means no one
uses them except for insane properties -- having +10 and a single property
for normal plate or +14 on CPM makes the latter a no-brainer once you have
some enchant scrolls.

This has a good chance of being too good, but let's give it a try.