Make Jiyva stat shuffling smarter.
The dependence on the last stat you increased was extremely non-obvious
without spoilers, and it also didn't play very well because it was too
inflexible. Now the bias for Jiyva stat shuffling just looks at your armour
and skills and tries to push your stats towards reasonable values based on
that information.