Rename "beast" to "hell beast".
We need monsters named "monster", "insect", "mammal", or "creature".
Unlike, say, "human" which is about as specific as "orc" or "yaktaur", or
"dragon" which refers to the ordinary species (like "bat" vs "fire bat"),
the name "beast" doesn't say anything about the monster.

The new name might be not the best choice, if you have a better idea, please
say so.