A vault showcasing tornado rotation.  The only way in is to let yourself get carried away.
This implementation isn't that good, though.  A single scroll of blinking and a means of
teleportation mean you can get the reward taking just one turn of damage.  I tried making
a whole tornado alley but I guess some people are better at vault making at me.  Another
idea would be a system (at least a loop) of tight corridors, damage is proportional to
obelisk hd so reduction is not an issue.