Make Tornado Conj/Air again.
With good apts and Ashenzari you can get it before Swamp/Snake, it's have
to either be nerfed to nothing or made harder (level 10, two school like
Firestorm, etc).

The alternative to Conj would be Tloc (translocating air to cause
differences in pressure, starting a wind system) but Conj is more
conservative and is supposed to have a near monopoly on direct damage.