Make show_map() more consistent.
Previously map-based targetting gave an advantage to players using larger
terminals in console (see Mantis 4180) and also to players playing tiles,
by allowing them to target more squares. In addition, it leaked information
to tiles players about their north/south position on the map.

This commit makes the set of map-targetable squares not depend on any of these
interface choices, which also removes the tiles information leak. The chosen
set is simply the smallest rectangle that contains every square visible
on the map.

This is a nerf to effects such as controlled teleports that used the map to
target, but you rarely wanted to make a controlled teleport outside of the
known parts of the map unless you had looked up a vault layout and counted
squares... which was not a good thing anyway.