* Implementable: launcher/weapon autoswitching
Switching from melee to ranged only considered
if ranged weapon is in slot A or B and the melee
weapon is in B or A.
Switching from ranged to melee picks first melee
weapon (behaviour is as above if A or B slot  contains
a melee weapon) Rationale is you don't lose much
by switching to any melee weapon with respect to ranged
one.
Added autoswitch option. Enabled by default.There was
a duplicate easy_open entry l_option.cc.

Signed-off-by: Raphael Langella <raphael.langella@gmail.com>