Fix the wizard mode piety gain command.
Before, the piety difference would be decreased by penance and gift
timeout, so you'd end up with less piety than you specified.  As a side
effect of the fix, piety no longer needs to be set to 0 afterward, as
it's now set to 0 as part of the original piety increase.