Drop tile monster cache whenever new monster/player tiles are added.
(Player tiles because of ghosts).

Sadly, this leads to loss of shown information since we currently can't
recreate the tiles from mon_info -- you need to get into view of that
monster again to correct them.

Once recreation is fixed, we may stop saving the mcache at all, it's
redundant with mon_info.  I'm adding this commit only for the short term,
in case we can't fix it properly in time for 0.9.