Make Berserkers start with a slightly more species-appropriate weapon
Starting everyone with an axe just results in most races dropping it
immediately for a spear or short sword as appropriate - it doesn't
really add much other than mild annoyance on D:1. Kobolds deliberately
get a mace over a short sword with this change, since that's a case
where either sticking with the mace or swapping to short/long blades
are both reasonable things to do.

Conflicts:
	crawl-ref/source/ng-setup.cc